MemoViewを廃止し、処理をMemoAppに集約
