+++ 04.class/lib/memo_app.js
@@ -1,24 +1,58 @@
+import enquirer from "enquirer";
 import memoStore from "./memo_store.js";
 import MemoEntry from "./memo_entry.js";
 
+const { Select } = enquirer;
+
 export default class MemoApp {
-  memos = [];
+  constructor(memos) {
+    this.memos = memos;
+  }
 
-  static async load() {
-    const app = new MemoApp();
+  static async build() {
     const memosData = await memoStore.load();
-    app.memos = MemoEntry.all(memosData);
-    return app;
+    const memos = MemoEntry.all(memosData);
+    return new MemoApp(memos);
   }
 
-  add(content) {
+  list() {
+    for (const memo of this.memos) {
+      console.log(memo.firstLine());
+    }
+  }
+
+  async reference() {
+    const prompt = this.#selectMemo("Select a memo to display.");
+    const memo = await prompt.run();
+    console.log(memo.content);
+  }
+
+  async delete() {
+    const prompt = this.#selectMemo("Select a memo to delete.");
+    const selectedMemo = await prompt.run();
+    this.memos = this.memos.filter((memo) => memo.id !== selectedMemo.id);
+    await memoStore.save(this.memos);
+  }
+
+  async add(content) {
     const memo = MemoEntry.create(content);
-    const newMemos = this.memos.concat(memo);
-    return memoStore.save(newMemos);
+    this.memos = this.memos.concat(memo);
+    await memoStore.save(this.memos);
   }
 
-  delete(id) {
-    const newMemos = this.memos.filter((memo) => memo.id !== id);
-    return memoStore.save(newMemos);
+  #selectMemo(message) {
+    const choices = this.memos.map((memo) => ({
+      name: memo.firstLine(),
+      value: memo,
+    }));
+
+    return new Select({
+      name: "memo",
+      message,
+      choices,
+      result() {
+        return this.focused.value;
+      },
+    });
   }
 }
